Update sp_selectpages.sql
CHANGE: An optional parameter that indicates that the result needs to be inserted in another table.
REASON: INSERT INTO EXEC(sp_select..) is not possible. "Nested inserts issue"
PROCEDURES CHANGED: sp_select and sp_selectpages

Hi, 

Thanks for the procedure. I'm using it a lot during debugs.  

Since the program generates a lot of tables, I'd like to store the result in another table to analyse it later. That's why I'm suggesting this change. There won't be any impact in other codes, because the parameter is optional and at the end.  

Best,
Eduardo